重构ast, 使添加原件轻量化
