+++ mdex/ast/sch.go
@@ -6,32 +6,22 @@ import (
 	"log"
 	"regexp"
 	"strconv"
-	"strings"
 
 	svg "github.com/ajstarks/svgo"
 	gast "github.com/yuin/goldmark/ast"
 )
 
-/*
-* 第一步先渲染一个芯片
-* $ U10-P8-NSTC12[1:VCC,8:GND] $
- */
+const (
+	div = 12 // 1div = 12 pt
+)
 
-// ICBlock 芯片描述
-type ICBlock struct {
-	ICIndex  string //U10
-	ICPins   int    //P8
-	ICName   string // NSTC
-	PinNames map[string]string
-	Y        int
-	X        int
-	W        int
-}
+// 保存解析器
+var schParsers = make([]Lineparser, 0, 10)
 
 // SchBlock 原理图
 type SchBlock struct {
 	gast.BaseBlock
-	Ics   []*ICBlock
+	Ics   []SvgBlock
 	PageW int
 	PageH int
 }
@@ -52,71 +42,35 @@ func (n *SchBlock) Kind() gast.NodeKind {
 	return KindSchBlock
 }
 
-// AddLine 添加一个行描述符
-func (n *SchBlock) AddLine(desc string) int {
-	log.Println(desc)
+// InitByLine 初始化画布
+func (n *SchBlock) InitByLine(desc string) {
 	// 读取画布大小
 	pageszre := regexp.MustCompile(`\$\(([0-9]+),([0-9]+)\)`)
 	pagesz := pageszre.FindStringSubmatch(desc)
 	if len(pagesz) == 3 {
 		n.PageW, _ = strconv.Atoi(pagesz[1])
 		n.PageH, _ = strconv.Atoi(pagesz[2])
-		return 0
+		return
 	}
-	// 拆出芯片编号->成功则添加一颗芯片
-	ux := regexp.MustCompile(`U([0-9]+)-`)
-	u := ux.FindStringSubmatch(desc)
-	if len(u) > 1 {
-		icb := new(ICBlock)
-		icb.PinNames = make(map[string]string)
-		n.Ics = append(n.Ics, icb)
+}
 
-		icb.ICIndex = u[1]
-		// log.Println(n.ICIndex, u[1])
-		desc = desc[len(u[0]):]
-		//拆出引脚数量
-		px := regexp.MustCompile(`P([0-9]+)-`)
-		p := px.FindStringSubmatch(desc)
-		if len(p) > 1 {
-			icb.ICPins, _ = strconv.Atoi(p[1])
-			desc = desc[len(p[0]):]
-		}
-		// 拆出芯片
-		nx := regexp.MustCompile(`N([A-Za-z0-9]+)`)
-		names := nx.FindStringSubmatch(desc)
-		if len(names) > 1 {
-			icb.ICName = names[1]
-			desc = desc[len(names[0]):]
-		}
-		// 如果有[]则解析引脚命名
-		nstart := strings.Index(desc, "[")
-		if nstart >= 0 {
-			nend := strings.Index(desc, "]")
-			pinstr := desc[nstart+1 : nend]
-			desc = desc[nend+1:]
-			pins := strings.Split(pinstr, ",")
-			for _, v := range pins {
-				apin := strings.Split(v, ":")
-				icb.PinNames[apin[0]] = apin[1]
+// AddLine 添加一个行描述符
+func (n *SchBlock) AddLine(desc string) int {
+	log.Println(desc)
+
+	for _, v := range schParsers {
+		if v.CanParse(desc) {
+			lp := v.ParseLine(desc)
+			if lp != nil {
+				n.Ics = append(n.Ics, lp)
 			}
 		}
-		// 拆出位置信息
-		lc := regexp.MustCompile(`\(([0-9]+),([0-9]+),([0-9]+)\)`)
-		lsl := lc.FindStringSubmatch(desc)
-		if len(lsl) > 3 {
-			icb.X, _ = strconv.Atoi(lsl[1])
-			icb.Y, _ = strconv.Atoi(lsl[2])
-			icb.W, _ = strconv.Atoi(lsl[3])
-			desc = desc[len(lsl[0]):]
-		}
 	}
-	//
 	return len(desc)
 }
 
 // ToSvg 输出svg
 func (n *SchBlock) ToSvg(w io.Writer) {
-	div := 12
 	width := n.PageW * div
 	height := n.PageH * div
 	canvas := svg.New(w)
@@ -133,42 +87,7 @@ func (n *SchBlock) ToSvg(w io.Writer) {
 	}
 
 	for _, v := range n.Ics {
-		// 画中间框
-		canvas.Rect(v.X*div, v.Y*div, v.W*div, v.ICPins*div/2+1*div, "fill:#cdcdcf;stroke:#737375;stroke-width:1pt;")
-
-		// 芯片编号
-		canvas.Text(v.X*div, v.Y*div-div/2, "U"+v.ICIndex)
-		// 画芯片引脚,先用左右方式
-		// 左侧
-		for i := 0; i < v.ICPins/2; i++ {
-			// 引脚编号
-			strpin := strconv.Itoa(i + 1)
-			canvas.Text((v.X-2)*div, (v.Y+1+i)*div, strpin, "font-size:"+strconv.Itoa(div)+"px;")
-			// 引脚名称
-			name, ok := v.PinNames[strpin]
-			if ok {
-				canvas.Text((v.X)*div+2, (v.Y+1+i)*div+div/3, name, "font-size:"+strconv.Itoa(div)+"px;")
-			}
-			// 引脚线
-			canvas.Line(v.X*div, (v.Y+1+i)*div, (v.X-2)*div, (v.Y+1+i)*div, "stroke:#737375;")
-		}
-		// 右侧
-		for i := 0; i < v.ICPins/2; i++ {
-			// U型编号
-			strpin := strconv.Itoa(v.ICPins - i)
-			canvas.Text((v.X+v.W+1)*div, (v.Y+1+i)*div, strpin, "font-size:"+strconv.Itoa(div)+"px;")
-			// 引脚名称
-			name, ok := v.PinNames[strpin]
-			if ok {
-				canvas.Text((v.X+v.W)*div, (v.Y+1+i)*div+div/3, name, "font-size:"+strconv.Itoa(div)+"px;text-anchor: end")
-			}
-			// 引脚线
-			canvas.Line(v.X*div+v.W*div, (v.Y+1+i)*div, (v.X+v.W+2)*div, (v.Y+1+i)*div, "stroke:#737375;")
-		}
-
-		// 芯片名称
-		canvas.Text(v.X*div, v.Y*div+v.ICPins*div/2+2*div, v.ICName, "font-size:"+strconv.Itoa(div)+"px;")
-
+		v.ToSvg(canvas, w)
 	}
 
 	// canvas.Text(width/2, height/2, "Hello, SVG", "text-anchor:middle;font-size:30px;fill:white")
